Extract private test event helper
